+++ packages/ui/src/modals/UpdateMembershipModal/UpdateMembershipFormModal.tsx
@@ -22,6 +22,7 @@ import { useObservable } from '../../hooks/useObservable'
 import { AvatarURISchema, HandleSchema } from '../../membership/data/validation'
 import { Row } from '../common'
 import { FormReducer, Nullable, UpdateMemberForm } from './types'
+import { changedOrNull, hasAnyEdits } from './utils'
 
 interface Props {
   onClose: () => void
@@ -38,35 +39,6 @@ const UpdateMemberSchema = Yup.object().shape({
 
 const updateReducer: FormReducer<UpdateMemberForm> = (state, action): UpdateMemberForm => {
   return { ...state, [action.type]: action.value }
-}
-
-export const hasAnyEdits = (formData: Record<string, any>, member: Record<string, any>) => {
-  return !!getChangedFields(formData, member).length
-}
-
-export const getChangedFields = (form: Record<string, any>, initial: Record<string, any>) => {
-  const changedFields = []
-
-  for (const key of Object.keys(form)) {
-    const memberValue = initial[key] || ''
-    const formValue = form[key]?.address ?? (form[key] || '')
-    if (memberValue !== formValue) {
-      changedFields.push(key)
-    }
-  }
-
-  return changedFields
-}
-
-export const changedOrNull = <T extends any>(form: Record<string, any>, initial: Record<string, any>): Nullable<T> => {
-  const changedFields = getChangedFields(form, initial)
-
-  return Object.entries(form).reduce((prev, [key, value]) => {
-    return {
-      ...prev,
-      [key]: changedFields.includes(key) ? value : null,
-    }
-  }, {} as Nullable<T>)
 }
 
 export const UpdateMembershipFormModal = ({ onClose, onSubmit, member }: Props) => {
